+++ packages/google_sign_in/example/android/app/src/main/java/io/flutter/plugins/googlesigninexample/MainActivity.java
@@ -5,10 +5,10 @@
 package io.flutter.plugins.googlesigninexample;
 
 import android.os.Bundle;
-import io.flutter.app.FlutterFragmentActivity;
+import io.flutter.app.FlutterActivity;
 import io.flutter.plugins.GeneratedPluginRegistrant;
 
-public class MainActivity extends FlutterFragmentActivity {
+public class MainActivity extends FlutterActivity {
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);